+++ tests/MWSServiceTest.php
@@ -2,6 +2,7 @@
 
 namespace Looxis\LaravelAmazonMWS\Tests;
 
+use Looxis\LaravelAmazonMWS\MWSFeeds;
 use Looxis\LaravelAmazonMWS\MWSOrders;
 use Looxis\LaravelAmazonMWS\MWSService;
 use Orchestra\Testbench\TestCase;
@@ -13,5 +14,12 @@ class MWSServiceTest extends TestCase
     {
         $mws = new MWSService;
         $this->assertInstanceOf(MWSOrders::class, $mws->orders());
+    }
+
+    /** @test */
+    public function get_feeds_service()
+    {
+        $mws = new MWSService;
+        $this->assertInstanceOf(MWSFeeds::class, $mws->feeds());
     }
 }